expect capabilities, and broadcast concrete thread count

+++ controller/src/message/capabilities.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 pub struct Capabilities {
     /// Maximum number of threads that the application can use.
     /// If `None`, thread adjustment is disabled.
-    pub max_threads: Option<usize>,
+    pub max_threads: Option<u16>,
 }
